+++ cookbooks/base/metadata.rb
@@ -4,6 +4,6 @@ maintainer_email 'carles.ml.dev@gmail.com'
 license          'Apache 2.0'
 description      'Installs/Configures base'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.27.1'
+version          '0.28.0'
 depends          'chef_handler'
 depends          'sudo'