+++ cookbooks/base/recipes/end-linuxmint.rb
@@ -21,52 +21,12 @@
 
 ## First time system upgrade
 
-held_pkgs = %w[
-  linux-generic
-  linux-headers-generic
-  linux-image-generic
-  base-files
-  desktop-file-utils
-  grub-common
-]
-
-upgrade_code = case platform_desktop
-when "cinnamon"
-  purge_pkgs = %w[
-    nautilus
-    nautilus-gksu
-    nautilus-open-terminal
-    nautilus-sendto-empathy
-    nautilus-wallpaper
-    gnome-session-fallback
-    gnome-control-center
-  ]
-
-  h_pkgs = held_pkgs.join(" ")
-  p_pkgs = purge_pkgs.join(" ")
-
-  <<-EOH
-    apt-mark hold #{h_pkgs}
-    apt-get -y dist-upgrade
-    apt-get -y purge #{p_pkgs}
-    apt-get -y autoremove
-    apt-get clean
-    apt-mark unhold #{h_pkgs}
-    EOH
-when "mate", "xfce"
-  h_pkgs = held_pkgs.join(" ")
-
-  <<-EOH
-    apt-mark hold #{h_pkgs}
-    apt-get -y dist-upgrade
-    apt-get -y autoremove
-    apt-get clean
-    apt-mark unhold #{h_pkgs}
-    EOH
-end
-
 bash "first_system_upgrade" do
-  code upgrade_code
+  code <<-EOH
+    apt-get -y upgrade
+    apt-get -y autoremove
+    apt-get clean
+    EOH
   action :run
   not_if { node.attribute?(:first_run_completed) }
 end